move SF helpers to separate file
