+++ packages/aws-client/step-functions.js
@@ -1,0 +1,98 @@
+const { JSONPath } = require('jsonpath-plus');
+const string = require('@cumulus/common/string');
+
+/**
+ * Given an array of fields, returns that a new string that's safe for use as a StepFunction,
+ * execution name, where all fields are joined by a StepFunction-safe delimiter
+ * Important: This transformation isn't entirely two-way. Names longer than 80 characters
+ *            will be truncated.
+ *
+ * @param {string} fields - The fields to be injected into an execution name
+ * @param {string} delimiter - An optional delimiter string to replace, pass null to make
+ *   no replacements
+ * @returns {string} A string that's safe to use as a StepFunctions execution name
+ */
+exports.toSfnExecutionName = (fields, delimiter = '__') => {
+  let sfnUnsafeChars = '[^\\w-=+_.]';
+  if (delimiter) {
+    sfnUnsafeChars = `(${delimiter}|${sfnUnsafeChars})`;
+  }
+  const regex = new RegExp(sfnUnsafeChars, 'g');
+  return fields.map((s) => s.replace(regex, string.unicodeEscape).replace(/\\/g, '!'))
+    .join(delimiter)
+    .substring(0, 80);
+};
+
+/**
+ * Opposite of toSfnExecutionName. Given a delimited StepFunction execution name, returns
+ * an array of its original fields
+ * Important: This value may be truncated from the original because of the 80-char limit on
+ *            execution names
+ *
+ * @param {string} str - The string to make stepfunction safe
+ * @param {string} [delimiter='__'] - An optional delimiter string to replace, pass null to make
+ *   no replacements
+ * @returns {Array} An array of the original fields
+ */
+exports.fromSfnExecutionName = (str, delimiter = '__') =>
+  str.split(delimiter)
+    .map((s) => s.replace(/!/g, '\\').replace('"', '\\"'))
+    .map((s) => JSON.parse(`"${s}"`));
+
+/**
+ * Returns execution ARN from a statement machine Arn and executionName
+ *
+ * @param {string} stateMachineArn - state machine ARN
+ * @param {string} executionName - state machine's execution name
+ * @returns {string} - Step Function Execution Arn
+ */
+exports.getExecutionArn = (stateMachineArn, executionName) => {
+  if (stateMachineArn && executionName) {
+    const sfArn = stateMachineArn.replace('stateMachine', 'execution');
+    return `${sfArn}:${executionName}`;
+  }
+  return null;
+};
+
+exports.getStateMachineArn = (executionArn) => {
+  if (executionArn) {
+    return executionArn.replace('execution', 'stateMachine').split(':').slice(0, -1).join(':');
+  }
+  return null;
+};
+
+/**
+ * Given a Cumulus step function event, if the message is on S3, pull the full message
+ * from S3 and return, otherwise return the event.
+ *
+ * @param {Object} event - the Cumulus event
+ * @returns {Object} - the full Cumulus message
+ */
+exports.pullStepFunctionEvent = async (event) => {
+  if (!event.replace) return event;
+
+  const remoteMsgS3Object = await exports.getS3Object(
+    event.replace.Bucket,
+    event.replace.Key,
+    { retries: 0 }
+  );
+  const remoteMsg = JSON.parse(remoteMsgS3Object.Body.toString());
+
+  let returnEvent = remoteMsg;
+  if (event.replace.TargetPath) {
+    const replaceNodeSearch = JSONPath({
+      path: event.replace.TargetPath,
+      json: event,
+      resultType: 'all'
+    });
+    if (replaceNodeSearch.length !== 1) {
+      throw new Error(`Replacement TargetPath ${event.replace.TargetPath} invalid`);
+    }
+    if (replaceNodeSearch[0].parent) {
+      replaceNodeSearch[0].parent[replaceNodeSearch[0].parentProperty] = remoteMsg;
+      returnEvent = event;
+      delete returnEvent.replace;
+    }
+  }
+  return returnEvent;
+};
